Add day 9.

+++ 2019/solutions/Day05_SunnyWithAChanceOfAsteroids.main.kts
@@ -5,17 +5,17 @@ class Day05_SunnyWithAChanceOfAsteroids {
     companion object {
 
         fun getAnswer(lines: List<String>, part: Int): Long {
-            val program = lines[0].split(",").map { it.toInt() }
+            val program = lines[0].split(",").map { it.toLong() }
 
-            val input = if (part == 1) 1 else 5
+            val input = if (part == 1) 1L else 5L
 
             val intcode = Intcode(program)
             intcode.pushInput(input)
-            var output: Int? = null
+            var output: Long? = null
             while (true) {
                 try {
                     val outputNext = intcode.run()
-                    if (output != null && output != 0) {
+                    if (output != null && output != 0L) {
                         throw RuntimeException("Unexpected output for instruction: $output")
                     }
                     output = outputNext
@@ -23,7 +23,7 @@ class Day05_SunnyWithAChanceOfAsteroids {
                     break
                 }
             }
-            return output!!.toLong()
+            return output!!
         }
     }
 }
